requirements.txt
absl-py==1.2.0
anyio==3.6.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
audioread==3.0.0
Babel==2.10.3
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.0.0
fastjsonschema==2.16.1
Flask==2.2.2
flatbuffers==1.12
fonttools==4.37.1
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.1
h5py==3.7.0
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.10
jsonschema==3.2.0
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyter_client==7.3.5
jupyterlab==3.4.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.15.1
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
librosa==0.9.2
llvmlite==0.39.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.3
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
notebook-shim==0.1.0
numba==0.56.2
numpy==1.23.2
oauthlib==3.2.0
opencv-python==4.4.0.40
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pooch==1.6.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.2.1
pyzmq==23.2.1
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
sniffio==1.3.0
soundfile==0.11.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.5.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.0
Werkzeug==2.2.2
wrapt==1.14.1
xgboost==1.6.2
zipp==3.8.1
